Exercise 1

+++ 05.fundamentalDatatypes/Exercise_1.cpp
@@ -11,8 +11,9 @@ int main(){
 
 	int days = age * 365;
 	int hours = age * 8760;
-	int minutes = age * 525600;
-	int seconds = age * 31540000;
+	long int minutes = age * 525600;
+	long long int seconds = age * 31540000;
+
 
 	cout<< "You are ... \n" << days << " days old, or \n" << hours << " hours old, or \n" << minutes << " minutes old, or \n" << seconds << " seconds old.\n";
